Rebase profile controller and page
